Add files via upload
Description:
This project is a simple Python script using the PyTube library to download YouTube videos. Users can input the YouTube video URL, and the script will download the video in the highest resolution available to a specified output directory. The script also checks for or creates a 'descargas' directory within the project folder to store the downloaded videos. It serves as a basic yet functional tool for easily downloading YouTube videos using Python.

Features:

Downloads YouTube videos in the highest resolution.
User-friendly input for the YouTube video URL.
Creates a 'descargas' directory for storing downloaded videos.
How to Use:

Run the script.
Input the YouTube video URL when prompted.
The script will download the video in the highest resolution to the 'descargas' directory.